♻️ refactor(mcp-local-server-unification): extract sandbox_tools.py, share across all servers
- run_python and run_style_python moved to sandbox_tools.py (not ACP-specific)
- All 3 servers now register sandbox tools + import_attachment
- server_acp.py reduced from 618 to 146 lines
- Only truly ACP-specific tool remaining: hearing (UI card display)

Tool counts: server.py=21, server_acp.py=22(+hearing), swi=22(+start_presentation)
